+++ tests/apps/wsgi/route.py
@@ -18,7 +18,7 @@ class Routes(unittest.TestCase):
         r = Route('bla/')
         self.assertFalse(r.is_leaf)
         self.assertEqual(r.level,1)
-        self.assertEqual(len(r.arguments),0)
+        self.assertEqual(len(r.variables), 0)
         self.assertEqual(r.rule,'bla/')
         self.assertEqual(r.match('bla/'),{})
         self.assertEqual(r.match('bladdd/'),None)
@@ -28,7 +28,7 @@ class Routes(unittest.TestCase):
         r = Route('bla/foo/')
         self.assertFalse(r.is_leaf)
         self.assertEqual(r.level,2)
-        self.assertEqual(len(r.arguments), 0)
+        self.assertEqual(len(r.variables), 0)
         self.assertEqual(r.rule, 'bla/foo/')
         self.assertEqual(r.match('bla/'), None)
         self.assertEqual(r.match('bla/foo/'), {})
@@ -56,7 +56,7 @@ class Routes(unittest.TestCase):
     def testStringVariable(self):
         r = Route('<name>/')
         self.assertFalse(r.is_leaf)
-        self.assertEqual(r.arguments,set(['name']))
+        self.assertEqual(r.variables, set(['name']))
         self.assertEqual(r.breadcrumbs,((True,'name'),))
         self.assertEqual(r.rule,'<name>/')
         self.assertEqual(r.match('bla-foo/'),{'name':'bla-foo'})
@@ -68,8 +68,8 @@ class Routes(unittest.TestCase):
         r = Route('<name>/<child>/')
         self.assertFalse(r.is_leaf)
         self.assertEqual(r.level,2)
-        self.assertEqual(r.arguments,set(['name','child']))
-        self.assertEqual(r.breadcrumbs,((True,'name'),(True,'child')))
+        self.assertEqual(r.variables, set(['name','child']))
+        self.assertEqual(r.breadcrumbs, ((True,'name'),(True,'child')))
         self.assertEqual(r.rule,'<name>/<child>/')
         self.assertEqual(r.match('bla/foo/'),{'name':'bla','child':'foo'})
         self.assertEqual(r.match('bla/foo/another/'),
@@ -91,7 +91,7 @@ class Routes(unittest.TestCase):
     def testIntVariable(self):
         r = Route('<int:id>/')
         self.assertEqual(str(r),'/<int:id>/')
-        self.assertEqual(r.arguments,set(['id']))
+        self.assertEqual(r.variables, set(['id']))
         self.assertEqual(r.breadcrumbs,((True,'id'),))
         self.assertEqual(r.match('35/'),{'id':35})
         self.assertEqual(r.url(id = 1), '/1/')
@@ -100,8 +100,8 @@ class Routes(unittest.TestCase):
     def testIntVariableFixDigits(self):
         r = Route('<int(2):id>/')
         self.assertEqual(str(r),'/<int(2):id>/')
-        self.assertEqual(r.arguments,set(['id']))
-        self.assertEqual(r.breadcrumbs,((True,'id'),))
+        self.assertEqual(r.variables, set(['id']))
+        self.assertEqual(r.breadcrumbs, ((True,'id'),))
         self.assertEqual(r.match('35/'),{'id':35})
         self.assertEqual(r.match('355/'),None)
         self.assertEqual(r.match('6/'),None)
@@ -115,7 +115,7 @@ class Routes(unittest.TestCase):
     def testIntVariableMinMax(self):
         r = Route('<int(min=1):cid>/')
         self.assertEqual(str(r),'/<int(min=1):cid>/')
-        self.assertEqual(r.arguments,set(['cid']))
+        self.assertEqual(r.variables, set(['cid']))
         self.assertEqual(r.breadcrumbs,((True,'cid'),))
         self.assertEqual(r.match('1/'),{'cid':1})
         self.assertEqual(r.match('476876/'),{'cid':476876})
@@ -129,7 +129,7 @@ class Routes(unittest.TestCase):
         
     def testPathVaiable(self):
         r = Route('bla/<path:rest>', defaults = {'rest':''})
-        self.assertEqual(r.arguments,set(['rest']))
+        self.assertEqual(r.variables, set(['rest']))
         self.assertEqual(r.level,2)
         self.assertTrue(r.is_leaf)
         self.assertEqual(r.match('bla/a/b/c.html'),{'rest':'a/b/c.html'})